+++ scripts/test/Selenium/Agent/AgentPreferences.t
@@ -107,13 +107,38 @@ $Selenium->RunTest(
 
         # check for some settings
         for my $ID (
-            qw(CurPw NewPw NewPw1 UserLanguage OutOfOfficeOn OutOfOfficeOff UserGoogleAuthenticatorSecretKey)
+            qw(CurPw NewPw NewPw1 UserTimeZone_Search UserLanguage_Search OutOfOfficeOn OutOfOfficeOff UserGoogleAuthenticatorSecretKey GenerateUserGoogleAuthenticatorSecretKey)
             )
         {
             my $Element = $Selenium->find_element( "#$ID", 'css' );
-            $Element->is_enabled();
-            $Element->is_displayed();
+
+            $Self->True(
+                $Element->is_enabled(),
+                "$ID is enabled."
+            );
+
+            $Self->True(
+                $Element->is_displayed(),
+                "$ID is displayed."
+            );
         }
+
+        # Click on "Generate" button.
+        $Selenium->find_element( "#GenerateUserGoogleAuthenticatorSecretKey", 'css' )->click();
+
+        # Wait until generated key is there.
+        $Selenium->WaitFor(
+            JavaScript =>
+                "return \$('#UserGoogleAuthenticatorSecretKey').val().length"
+        );
+
+        my $SecretKey = $Selenium->execute_script(
+            "return \$('#UserGoogleAuthenticatorSecretKey').val();"
+        );
+        $Self->True(
+            $SecretKey =~ m{[A-Z2-7]{16}} ? 1 : 0,
+            'Secret key is valid.'
+        );
 
         # check some of AgentPreferences default values
         $Self->Is(
@@ -459,6 +484,8 @@ JAVASCRIPT
             "ivory",
             "#UserSkin updated value",
         );
+
+        # Enable two factor authenticator.
 
         if ( $Kernel::OM->Get('Kernel::System::OTRSBusiness')->OTRSBusinessIsInstalled() ) {